Refactor level creation to utils module
Put height level creation into global `utils` module so that it can be used in both forcing calculation and conversions. Also makes it easier to use lagtraj to simply convert era5 data to height levels
