+++ lagtraj/utils/interpolation/methods.py
